requirements.txt
asgiref==3.11.1
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sqlparse==0.5.5
